Break dependency on mirrors.
This is a lot of changes and some are fundamental towards the way
that steward works.

- Remove support for Containers
- Adjust response type to be typesafe.
- Adjust examples
- View template rendering logic has been moved

This is a pretty critical commit, but ultimately we're increasing
typesafety, focusing on speed (reflection is somwhat slow) and
are allowing people to compile steward in AoT compilation (reflection
was preventing this). The trade off is the "magic" behind controllers
has been removed and a few new methods have been introduced.